# [8.4.0] Avoid `ByteString` allocation during zstd decompression