… src/java_opt_binary.bzl